+++ corner_detectors.h
@@ -15,19 +15,19 @@ using namespace std::placeholders;
 struct Point{
     int x;
     int y;
-    float value;
-    Point(const int x,const int y,const float v):x(x),y(y),value(v){};
+    double value;
+    Point(const int x,const int y,const double v):x(x),y(y),value(v){};
 };
 
 enum Algorithm {MORAVEC, HARIS};
 
 class CornerDetectors
 {
-    const float E = 1e-5;
-    const float sigma = 1;
+    const double E = 1e-5;
+    const double sigma = 1;
 
-    const float trasholdMor = 0.05;
-    const float trasholdHar = 0.25;
+    const double trasholdMor = 0.05;
+    const double trasholdHar = 0.25;
     const int winSize = 3;
     const int pSize = 3;
 
@@ -38,12 +38,12 @@ public:
 private:
     Matrix detectMoravec(const Matrix & m) const;
     Matrix detectHaris(const Matrix & m) const;
-    vector<Point> localMinimums(const Matrix & m, const float tr) const;
-    bool isMinimum(const Matrix & m,const int i,const int j, const float tr) const;
+    vector<Point> localMinimums(const Matrix & m, const double tr) const;
+    bool isMinimum(const Matrix & m,const int i,const int j, const double tr) const;
 };
 
 class PointFileter{
-    const float factor = 0.9;
+    const double factor = 0.9;
     int maxR;
     int maxPoints;
