+++ kernel.h
@@ -13,12 +13,12 @@ struct Kernel
 public:
     int width;
     int height;
-    unique_ptr<float []> matrix;
+    unique_ptr<double []> matrix;
 
 public:
     Kernel();
     Kernel(int width,int height);
-    Kernel(int width,int height,unique_ptr<float []>);
+    Kernel(int width,int height,unique_ptr<double []>);
     Kernel(Kernel &&) = default;
 };
 
@@ -28,9 +28,9 @@ public:
 class KernelFactory
 {
 public:
-    static Kernel createGauss(float sigma);
-    static Kernel createGaussX(float sigma);
-    static Kernel createGaussY(float sigma);
+    static Kernel createGauss(double sigma);
+    static Kernel createGaussX(double sigma);
+    static Kernel createGaussY(double sigma);
     static Kernel sobelX();
     static Kernel sobelY();
 };
