+++ matrix.cpp
@@ -3,7 +3,7 @@
 Matrix::Matrix():_width(0),_height(0),matrix(nullptr){}
 
 Matrix::Matrix(const int width,const int height):
-    _width(width),_height(height),matrix(make_unique<float[]>(_width*_height)){}
+    _width(width),_height(height),matrix(make_unique<double[]>(_width*_height)){}
 
 int Matrix::width() const{
     return _width;
@@ -13,7 +13,7 @@ int Matrix::height() const{
     return _height;
 }
 
-float Matrix::get(const int i,const int j, const Border border) const
+double Matrix::get(const int i,const int j, const Border border) const
 {
     switch (border) {
     case SIMPLE:
@@ -30,7 +30,7 @@ float Matrix::get(const int i,const int j, const Border border) const
     }
 }
 
-void Matrix::set(const int i,const int j,const float gray)
+void Matrix::set(const int i,const int j,const double gray)
 {
     matrix[i*_height + j] = gray;
 }
@@ -38,7 +38,7 @@ void Matrix::set(const int i,const int j,const float gray)
 Matrix Matrix::normalize() const
 {
     auto minMax = minmax_element(&matrix[0],&matrix[_width*_height]);
-    float range = (*minMax.second - *minMax.first);
+    double range = (*minMax.second - *minMax.first);
     auto funk = Utils::normalize(*minMax.first,range);
     return compute(funk);
 }
@@ -59,7 +59,7 @@ Matrix Matrix::compress() const
     Matrix compresed(_width/2,_height/2);
     for(int i = 0; i < compresed._width; i++){
         for(int j = 0; j < compresed._height; j++){
-            float average = (get(i*2,j*2,CILINDER)+
+            double average = (get(i*2,j*2,CILINDER)+
                             get(i*2+1,j*2,CILINDER)+
                             get(i*2,j*2+1,CILINDER)+
                             get(i*2+1,j*2+1,CILINDER))/4.0;
@@ -69,10 +69,10 @@ Matrix Matrix::compress() const
     return compresed;
 }
 
-float Matrix::convoluite(const int elI, const int elJ, const Kernel &k,
+double Matrix::convoluite(const int elI, const int elJ, const Kernel &k,
                          const Border border) const
 {
-    float brightness = 0;
+    double brightness = 0;
     for(int i = 0; i < k.width; i++){
         for(int j = 0; j < k.height; j++){
             int posX = elI + (i - (k.width / 2));
@@ -83,7 +83,7 @@ float Matrix::convoluite(const int elI, const int elJ, const Kernel &k,
     return brightness;
 }
 
-Matrix Matrix::compute(function<float (float)> &funk) const
+Matrix Matrix::compute(function<double (double)> &funk) const
 {
     Matrix computed(_width,_height);
     transform(matrix.get(),matrix.get()+_width*_height,computed.matrix.get(),funk);
@@ -91,7 +91,7 @@ Matrix Matrix::compute(function<float (float)> &funk) const
 }
 
 Matrix Matrix::compute(const Matrix &other,
-                       const function<float (float, float)> & funk) const
+                       const function<double (double, double)> & funk) const
 {
     assert(_width == other._width && _height == other._height);
     Matrix computed(_width,_height);
