+++ scalespace.h
@@ -13,18 +13,18 @@ class ScaleSpace
 public:
     struct ScaleLevel{
         Matrix matrix;
-        float sigma;
-        float efectSigma;
+        double sigma;
+        double efectSigma;
         ScaleLevel():sigma(0),efectSigma(0),matrix(Matrix(0,0)){}
         ScaleLevel(ScaleLevel &&) = default;
-        ScaleLevel(Matrix & m,float sigma,float efectSigma):
+        ScaleLevel(Matrix & m,double sigma,double efectSigma):
             sigma(sigma),efectSigma(efectSigma),matrix(move(m)){}
     };
 private:
     const int _layerSize = 7;
-    const float startSigma = 0.5;
-    const float sigmaA = 1.6;
-    const float k = pow(2,1/(float)_layerSize);
+    const double startSigma = 0.5;
+    const double sigmaA = 1.6;
+    const double k = pow(2,1/(double)_layerSize);
 
     vector<vector<ScaleLevel>> _octavs;
 
@@ -36,7 +36,7 @@ public:
     const vector<vector<ScaleLevel>> & octavs() const;
 
 private:
-    Matrix gauss(const Matrix & matrix,const float sigma) const;
+    Matrix gauss(const Matrix & matrix,const double sigma) const;
     void calculate(const Matrix & m);
 };
